+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed getting recipient options in user profile chat
+- Fix translation key for 'Date' in view_snapshots.html
 
 ## [2.3.0] - 2026-04-19