+++ CHANGELOG.md
@@ -13,7 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed getting recipient options in user profile chat
-- Fix translation key for 'Date' in view_snapshots.html
+- Fixed translation key for 'Date' in view_snapshots.html
+
+# Improved
+
+- Enhanced TaskAdmin: update field ordering and add custom form for task widget
 
 ## [2.3.0] - 2026-04-19